+++ packages/amplify-gen2-codegen/src/resource/resource.test.ts
@@ -8,13 +8,14 @@ const factory = ts.factory;
 describe('Resource.ts file generation', () => {
   describe('imports', () => {
     const importedFunctionName = 'helloWorld';
-    const importedPackageName = 'my-hello-world-package';
+    const additionalImportedBackendIdentifiers: Record<string, Set<string>> = { 'my-hello-world-package': new Set() };
+    additionalImportedBackendIdentifiers['my-hello-world-package'].add(importedFunctionName);
     const exportedVariableName = 'goodbyeWorld';
     const render = (parameters?: Partial<ResourceTsParameters>) =>
       printNodeArray(
         renderResourceTsFile({
           backendFunctionConstruct: importedFunctionName,
-          importedPackageName,
+          additionalImportedBackendIdentifiers,
           functionCallParameter: factory.createObjectLiteralExpression(),
           exportedVariableName: factory.createIdentifier(exportedVariableName),
           ...parameters,
@@ -25,13 +26,14 @@ describe('Resource.ts file generation', () => {
     });
     it('calls import with additionally import identifiers', () => {
       const additionalImport = 'aGoodDayToYou';
+      additionalImportedBackendIdentifiers['my-hello-world-package'].add(additionalImport);
       assert.match(
-        render({ additionalImportedBackendIdentifiers: [additionalImport] }),
+        render({ additionalImportedBackendIdentifiers }),
         new RegExp(`import \\{ ${importedFunctionName}, ${additionalImport} \\} from `),
       );
     });
     it('calls import with the correct package name', () => {
-      assert.match(render(), new RegExp(`from "${importedPackageName}";`));
+      assert.match(render(), new RegExp('from\\s+["\']my-hello-world-package["\']'));
     });
     it('makes the function call', () => {
       assert.match(render(), new RegExp(`${importedFunctionName}\\(\\{\\}\\);`));