+++ packages/amplify-app/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.37-gen2-migrations-test.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.36...@aws-amplify/amplify-app@5.0.37-gen2-migrations-test.0) (2024-09-23)
+
+**Note:** Version bump only for package @aws-amplify/amplify-app
+
+
+
+
+
 ## [5.0.36](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.35...@aws-amplify/amplify-app@5.0.36) (2024-08-02)
 
 **Note:** Version bump only for package @aws-amplify/amplify-app