+++ packages/amplify-category-geo/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.16-gen2-migrations-test.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-geo@3.5.15...@aws-amplify/amplify-category-geo@3.5.16-gen2-migrations-test.0) (2024-09-23)
+
+**Note:** Version bump only for package @aws-amplify/amplify-category-geo
+
+
+
+
+
 ## [3.5.15](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-geo@3.5.14...@aws-amplify/amplify-category-geo@3.5.15) (2024-08-02)
 
 **Note:** Version bump only for package @aws-amplify/amplify-category-geo